Add MovieController
